+++ src/Input/styled.ts
@@ -1,28 +1,26 @@
 import styled from 'styled-components'
 import { gray } from '../colors'
 
-const SEARCH_BOX_HEIGHT = '50px'
+const SEARCH_BOX_HEIGHT = '40px'
 
 export const Container = styled.div`
 	width: 100%;
 	height: ${SEARCH_BOX_HEIGHT};
 	box-sizing: border-box;
-	padding: 12px 16px;
-	box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.25);
-	border-radius: 16px;
+	padding: 8px 16px;
+	background: ${gray[0]};
+	border-radius: 100vw;
 	display: flex;
 	align-items: center;
-	background: white;
 	input {
 		outline: none;
 		border: none;
 		font-size: 18px;
+		font-weight: 300;
 		width: 100%;
+		background: ${gray[0]};
 		::placeholder {
 			color: ${gray[2]};
 		}
-	}
-	> *:not(:last-child) {
-		margin-right: 8px;
 	}
 `