+++ src/index.tsx
@@ -23,3 +23,5 @@ export { default as Gap } from './Gap'
 export { default as InputField } from './InputField'
 
 export { default as TextArea } from './TextArea'
+
+export { default as ImageUpload } from './ImageUpload'