+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.6](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.5...v0.1.6) (2019-09-13)
+
+
+### Bug Fixes
+
+* Correct Button Props ([b5e9f44](https://github.com/S-O-L-A-R/solor-ui/commit/b5e9f44))
+
+### [0.1.5](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.4...v0.1.5) (2019-09-13)
+
+
+### Bug Fixes
+
+* Grayscale invalid z-index ([f7917ac](https://github.com/S-O-L-A-R/solor-ui/commit/f7917ac))
+
+### [0.1.4](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.3...v0.1.4) (2019-09-13)
+
+
+### Bug Fixes
+
+* Add white background to SearchInput  ([67a5823](https://github.com/S-O-L-A-R/solor-ui/commit/67a5823))
+* Card Size to unset ([316fcab](https://github.com/S-O-L-A-R/solor-ui/commit/316fcab))
+
+### [0.1.3](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.1...v0.1.3) (2019-09-13)
+
+
+### Features
+
+* grayscale class ([6e85bb8](https://github.com/S-O-L-A-R/solor-ui/commit/6e85bb8))
+
+### [0.1.2](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.1...v0.1.2) (2019-09-13)
+
 ### [0.1.1](https://github.com/S-O-L-A-R/solor-ui/compare/v0.1.0...v0.1.1) (2019-09-13)
 
 ## 0.1.0 (2019-09-13)